+++ finagle-base-http/src/main/scala/com/twitter/finagle/http/Message.scala
@@ -6,7 +6,7 @@ import com.twitter.io.{Buf, BufInputStream, Reader, Writer}
 import com.twitter.util.{Duration, Future}
 import java.nio.charset.{Charset, StandardCharsets}
 import java.time.format.DateTimeFormatter
-import java.time.{ZoneId, ZoneOffset}
+import java.time.{Instant, ZoneId, ZoneOffset}
 import java.util.{Date, Locale, Iterator => JIterator}
 import scala.collection.JavaConverters._
 
@@ -592,7 +592,27 @@ object Message {
       .withLocale(Locale.ENGLISH)
       .withZone(ZoneId.of("GMT"))
 
+  /**
+   * Convert a [[java.util.Date]] into an RFC 7231 formatted String representation.
+   *
+   * @param date the [[java.util.Date]] to format.
+   * @return an RFC 7231 formatted String representation of the time represented by the given [[java.util.Date]].
+   *
+   * @see [[http://tools.ietf.org/html/rfc7231#section-7.1.1.1 RFC 7231 Section 7.1.1.1]]
+   */
   def httpDateFormat(date: Date): String = {
     date.toInstant.atOffset(ZoneOffset.UTC).format(HttpDateFormat)
+  }
+
+  /**
+   * Convert epoch milliseconds into an RFC 7231 formatted String representation.
+   *
+   * @param millis the milliseconds to format.
+   * @return an RFC 7231 formatted String representation of the time represented by the given milliseconds.
+   *
+   * @see [[http://tools.ietf.org/html/rfc7231#section-7.1.1.1 RFC 7231 Section 7.1.1.1]]
+   */
+  def httpDateFormat(millis: Long): String = {
+    HttpDateFormat.format(Instant.ofEpochMilli(millis))
   }
 }